requirements.txt
flask
torch
transformers
Pillow
requests
waitress
